chore: rename GlobalStyleProvider filename
