+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import 'ui/theme/classNameSetupOnImport';
 
-import { GlobalStyleProvider } from 'ui/theme/GlobalStyle';
+import { GlobalStyleProvider } from 'ui/theme/GlobalStyleProvider';
 import { AppProps } from 'next/app';
 import type { ReactElement, ReactNode } from 'react';
 import type { NextPage } from 'next';
